feat: Add breadcrumb component for navigation
This commit introduces a new breadcrumb component that aids in navigation. It uses the 'next/link' and 'next/navigation' libraries to get the current route and split it into segments. These segments are then used to create a breadcrumb trail for the user. The breadcrumb always starts with a link to the home page.